Added check to mic purge to make sure sampler contains multi-mic samples

+++ modules/genericMicMixer.js
@@ -172,7 +172,7 @@ function onControl(number, value)
 		{
 			for (s in samplers) //Each sampler
 			{
-				if (s.isMicPositionPurged(i) != value) //Only purge or load if it's not already purged or loaded
+				if (s.getNumMicPositions > 1 && s.isMicPositionPurged(i) != value) //Only purge or load if it's not already purged or loaded
 				{
 					s.purgeMicPosition(s.getMicPositionName(i), value);
 				}
